Don't actually build a device path
As it turns out, the version of libusb1.0 installed everywhere doesn't
support the "get port number" api calls.

And neither does dfu-util, which is what we're going to be flashing this
with anyway.

To work around this, progress with the assumption that there's only ever
one servo or power board attached. We can work out a superior solution
in the future.