Comment-out the power board entry in config.yaml
We currently have some teams with one power board firmware, and others
with another.  Since all teams now have the firmware they want, to
simplify matters, with this change we can ship the same image to all
teams.  We can also do firmware updates when we are at the event from
our laptops without the robot switching it back.

Change-Id: I9a6d8ad29ec744b618a7db4d1201fe749d91a15b